! Checkpoint. Test 2D preprocess against 1D.
- Change the initial conditions ot old horizontal network
to check that if the 2D code
is used to solve a 1D problem, then it does it correctly.
The results are the same.
- This is therefore a checkpoint for correct 2D implementation
of preprocessing.
- Future commits will use 2D code to recreate
result from Jon's paper.
- Future commits will also implement 2D flow.
Note that this might not be necessary as only interested in
flow in horizontal direction. So simply caculate the
zero component using the zero-zero permeability,
which is just the 1D flow problem.