Add plots to  permdist experiment
- Now plot script plots one graph for N=1 case, and
one for all N.
- The N=1 grpah compares histogram to the known distribution and
shows they are similar as num reps tends to infinity.
- The second graoh shows the histogram for each N on top of each other.
This might get busy and so consider taking a subset of n.

Future commits
- Plot the mean and standard deviation of the histograms
as a function of N. Compare this plot
to Jon's paper and use this to determine whether this code is right.
- If this is same as paper then code up to here is correct
and can be used for dynamic experiments.