Correct random conductance experiment
- This commit corrects random
conductance experiment
that results in distributions for k and j against N.
- After this commit, we run trial on
maths machines to see if now correct.
- OUTSTANDING ISSUE:
Have removed division by sqrt(N) and this
seems to give correct results... need to work out why division not necessary.
- Two  issues found from last run:
1. Structure of internal adjacency matrix for cond was wrong, i.e. wrong
connectivity.
2. Division by sqrt(N) happened
inside a loop which meant that
division happened to the order of
the loop length,
which pushed distributions
to zero.

Files changed:
cells.py:
- Add a class for creating a 2D grid inside a module designed for creating
cells.
- I have previosuly implemented this in the network models package, so
borrowed that class and edited.
- Future commits might integrate the two.

configure.py
- Uncomment random conductance inial
condition, add prescriebd, and comment
that and the random
structure initial conditions.
- Future commits will make everything an option in the config file,
so that this only ever needs to be edited to run.

initial_conditions_2D.py
- Make internal adjacency matric with
own class, rather than with networkx.
It looks like networkx indexing difference was causing problem.
- Fill internal adjacency matric with random numbers in new way.
- Take division by sqrt(N) outside loop.
- Remove division by sqrt(N) as it seems to give correct result.
This is an artedfact of the fact that larger cells
with ones on all edges still give permeability=1, in the same way that
smaller cells do.
This is unintuitive and we are probably not dividing
by the cell size in the right way.
- Future commits will understand this problem, however, removing division
provides an inintutive solution.

run_experiment_permdist.py
- This now saves the depodist as welll as the permdist.
- Future commits will clean up the name of this file, since
it is now more general than a permdist calculation.