Add plot code for k and j with N and l
- These codes plot effective permeability and
adhesivity against number of nodes
in cell and length of cell
in the case where we have a square lattice with
edges of conductance 1.
- In this case the results are explicit,
and we check them numerically with scatter points
in codes plot_kj__vs_N and plot_kj__vs__l.
- Then plot against N and sweep l
to illustrate k=N^(D-2/D)*l^(2-D)
and j=N^(D-1/D)*l
for N and l varying together.