Generalise heax cell runing and plotting
run_cell_hexag-struc.py
- Import configuration to cell run and
use this to generate cell with randomly sampled
conductances.

plot_cell_hexag-struc.py
- Generalise plots so that they agree with the
configuration.
This ensures that we are seeing what we
are actually running.

- Future commits might clean up
cell names and initial conditiiobn names,
so that they describe the standard or
random structure and then describe what distribution
conductances are drawn from or whether they are prescribed.