Major changes to cells module
More cell struictures and count the number of internal
and external cells.

- Create 6-ireglikereg: This forms the structure of an iregular cell
has the conductance of regular cells.

- create 6-reglikeireg: has structure of reg but conductance of ireg..,
note that this doesn't really have
conductance of ireg
because of the way we sample,
so shouldn't use this.

- Create 6-rand: Create cell that has truly random structure with connectivity
that is drawn from a distribution, and
conductance is the inversely proportional
to the length.