Make files for running network ensemble
run_ensemble-network
- bash script that loops runs of the network model
returns each solution to a different subdirectory.

run_ensemble-network-stats
- python script that collects solutions from subdirectories, calculates
the mean and standard deviation of the solutions,
and saves these in a new subdirectory.

plot_ensemble-network
- python script that plots the mean of the results with shaded
areas representing the standard deviation.

run_network
- made suitable for external control by bash script.
- new boolean: ensemble controls
whether this will be run as an ensemble file or as a single run.
- If ensemble then need to supply a path for the results to the bash file.
- Also add new parameter: incr which
is the increment of time at whcih solution is saved
as opposed to used in the run.