Remove depricated parameters
- Remove alpha and beta
Define adhe and effe distrubutions.
Parameter of adhe is alpha when
distribution is set to delta.
Similarly, parameter of effe is beta when
distribution is set to delta.
- Future commits consider overwrite is alpha and beta are supplied.

- num_dims
Only 2 is implemented so remove option.

- num_refs
Only 3 is implemented. Implement calculate num_refs based on max_ref_dist supplied.
Only 1 is implemented to raise Exception if other is given.

- v
Depricated parameter.

- Add leng_1 as parameter and remove from cell. This is property of cell
but considering in parameters removes need to call cell in
equations_preprocess class.